Fixed swerve (again), fixed pathplanner (again)
Implemented 364's permanent fix for bad swerve module angle resetting.
Added a more explicit pose resetting for pathplanner, as the built in one did not seem to reset properly. There are still issues with the way pathplanner rotates the robot back and forth, and spinning continously at the end of paths, but other than that it seems to work well
Added Test4 path, which does a basic run out of the community, and onto the charging station (kind of)
Added Pathplanner server start code