Working auto selector, added MoveChassis command
An auto selector is working, and uses enums for the different autos
MoveChassis command to send the chassis to a pose using PID, untested. Was created to "snap" to rotation using buttons but won't work for that purpose because it stops all other movement. We will need another solution for that. Deemed useful enough to keep the command though